fix: member 관련 api를 각 파일로 분리
