+++ src/entities/member/api/logout.ts
@@ -1,0 +1,5 @@
+import apiClient from '@/shared/api/apiClient';
+
+export const logout = async (): Promise<void> => {
+  await apiClient.post(`/auth/logout`, {}, { withCredentials: true });
+};
